server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django>=4.2
djangorestframework==3.15.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
frozenlist==1.4.1
h11==0.14.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uWSGI==2.0.24
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
yarl==1.9.4
